# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FlowLink API",
    "description" : "Project and portfolio management with a resource-constrained scheduling optimizer.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/users/me" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "getCurrentUser",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrentUserDTO"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "user-controller" ],
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "profile" : {
                    "type" : "string"
                  },
                  "profilePicture" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "profile" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrentUserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectKey}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProject",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "project-controller" ],
        "operationId" : "updateProject",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "projectDTO" : {
                    "type" : "string"
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "projectDTO" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "deleteProject",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/projects/{projectKey}/tasks/{taskKey}" : {
      "put" : {
        "tags" : [ "task-controller" ],
        "operationId" : "updateTask",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taskKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "taskDTO" : {
                    "type" : "string"
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "taskDTO" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "task-controller" ],
        "operationId" : "deleteTask",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taskKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/users" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRegistrationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjects",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseProjectDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "createProject",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "projectDTO" : {
                    "type" : "string"
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "projectDTO" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectKey}/tasks" : {
      "post" : {
        "tags" : [ "task-controller" ],
        "operationId" : "createTask",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "taskDTO" : {
                    "type" : "string"
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "taskDTO" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/email-preferences" : {
      "patch" : {
        "tags" : [ "user-controller" ],
        "operationId" : "updateEmailPreferences",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailPreferencesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrentUserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectKey}/tasks/{taskKey}/schedule" : {
      "patch" : {
        "tags" : [ "task-controller" ],
        "operationId" : "updateSchedule",
        "parameters" : [ {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taskKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskScheduleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/lookup" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "lookupByEmail",
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "CurrentUserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "profilePicture" : {
            "type" : "string"
          },
          "emailNotificationsEnabled" : {
            "type" : "boolean"
          },
          "emailOnTaskAssigned" : {
            "type" : "boolean"
          },
          "emailOnCommentReply" : {
            "type" : "boolean"
          },
          "emailOnProjectInvitation" : {
            "type" : "boolean"
          }
        }
      },
...
      "ProjectDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "projectKey" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "tasks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaskDTO"
            }
          },
          "members" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserDTO"
            }
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "owner" : {
            "$ref" : "#/components/schemas/UserDTO"
          },
          "dependencies" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "created" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updated" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "TaskDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taskNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taskKey" : {
            "type" : "string"
          },
          "projectKey" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "BACKLOG", "GATHERING_INTEREST", "TODO", "WITHDRAWN", "IN_PROGRESS", "TO_REVIEW", "TO_TEST", "IN_TEST", "READY_TO_MERGE", "READY_TO_DEPLOY", "RELEASED", "DONE" ]
          },
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "assignee" : {
            "type" : "string"
          },
          "labels" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "dependencyKeys" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "isCritical" : {
            "type" : "boolean"
          },
          "totalFloat" : {
            "type" : "integer",
            "format" : "int32"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "created" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "progress" : {
            "type" : "integer",
            "format" : "int32"
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "LOWEST", "LOW", "MEDIUM", "HIGH", "HIGHEST" ]
          }
        }
      },
      "UserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "profilePicture" : {
            "type" : "string"
          }
        }
      },
      "UserRegistrationRequest" : {
        "type" : "object",
        "properties" : {
          "firstname" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 2
          },
          "lastname" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 2
          },
          "email" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 8
          }
        }
      },
      "LoginResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "EmailPreferencesRequest" : {
        "type" : "object",
        "properties" : {
          "emailNotificationsEnabled" : {
            "type" : "boolean"
          },
          "emailOnTaskAssigned" : {
            "type" : "boolean"
          },
          "emailOnCommentReply" : {
            "type" : "boolean"
          },
          "emailOnProjectInvitation" : {
            "type" : "boolean"
          }
        }
      },
      "TaskScheduleRequest" : {
        "type" : "object",
        "properties" : {
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "dueDate", "startDate" ]
      },
...
      "PagedResponseProjectDTO" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProjectDTO"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}